+++ platform-training-data/training-data-codeforces/src/main/java/com/custacm/platform/trainingdata/codeforces/domain/repo/CodeforcesSubmissionRepository.java
@@ -1,0 +1,13 @@
+package com.custacm.platform.trainingdata.codeforces.domain.repo;
+
+import com.custacm.platform.trainingdata.codeforces.domain.criteria.CodeforcesHandleSubmissionCriteria;
+import com.custacm.platform.trainingdata.codeforces.domain.criteria.CodeforcesProblemSubmissionCriteria;
+import com.custacm.platform.trainingdata.codeforces.domain.model.CodeforcesSubmission;
+
+import java.util.List;
+
+public interface CodeforcesSubmissionRepository {
+    List<CodeforcesSubmission> findHandleSubmissions(CodeforcesHandleSubmissionCriteria query);
+
+    List<CodeforcesSubmission> findProblemSubmissions(CodeforcesProblemSubmissionCriteria query);
+}